Add Clear directory option
git-svn-id: https://scm.gforge.inria.fr/anonscm/svn/kermeta/trunk@12826 3b3ea266-f0f1-470e-95cb-5b2c9abc2e95

Former-commit-id: d33767af14707d337e8a8bc063e245cd9c83a1e5

+++ maventools_projects/org.kermeta.emf.genmodel/src/main/java/org/kermeta/emf/genmodel/Util.java
@@ -50,7 +50,7 @@ public class Util {
         return (path.delete());
     }
 
-    public static void createGenModel(File ecore, File genmodel, File sourcePath, Log log) {
+    public static void createGenModel(File ecore, File genmodel, File sourcePath, Log log,Boolean bool) {
         Resource.Factory.Registry.INSTANCE.getExtensionToFactoryMap().
                 put("ecore", new org.eclipse.emf.ecore.xmi.impl.EcoreResourceFactoryImpl());
         Resource.Factory.Registry.INSTANCE.getExtensionToFactoryMap().
@@ -98,7 +98,10 @@ public class Util {
             }
         }
 
-        deleteDirectory(sourcePath);
+        if(bool) {
+            log.info("Clear output directory , "+sourcePath.getAbsolutePath());
+            deleteDirectory(sourcePath);
+        }
         sourcePath.mkdir();
         EcorePlugin.getPlatformResourceMap().put(outputfileName, URI.createFileURI(sourcePath.getAbsolutePath() + "/"));
         Util.generate(genModelModel, log);
